chore: move sdk examples to its own folder
