+++ example/src/App.tsx
@@ -8,17 +8,17 @@ import SitumPlugin from '@situm/react-native';
 import {SITUM_API_KEY, SITUM_DASHBOARD_URL} from './situm';
 import Theme from './examples/styles/theme';
 
-import PositioningScreen from './examples/Positioning';
-import {BuildingsBasicInfo} from './examples/BuildingsBasicInfo';
-import {BuildingFullInfo} from './examples/BuildingFullInfo';
-import {InfoFromBuilding} from './examples/InfoFromBuilding';
-import {RouteBetweenPOIs} from './examples/RouteBetweenPOIs';
-import {ShowBuildingOnMap} from './examples/ShowBuildingOnMap';
-import {DrawRouteBetweenPOIs} from './examples/DrawRouteBetweenPOIs';
-import {RemoteConfig} from './examples/RemoteConfig';
-import {GetPoisIcons} from './examples/GetPoisIcons';
-import {SetCacheMaxAge} from './examples/SetCacheMaxAge';
-import {TiledBuilding} from './examples/TiledBuilding';
+import PositioningScreen from './examples/sdk/Positioning';
+import {BuildingsBasicInfo} from './examples/sdk/BuildingsBasicInfo';
+import {BuildingFullInfo} from './examples/sdk/BuildingFullInfo';
+import {InfoFromBuilding} from './examples/sdk/InfoFromBuilding';
+import {RouteBetweenPOIs} from './examples/sdk/RouteBetweenPOIs';
+import {ShowBuildingOnMap} from './examples/sdk/ShowBuildingOnMap';
+import {DrawRouteBetweenPOIs} from './examples/sdk/DrawRouteBetweenPOIs';
+import {RemoteConfig} from './examples/sdk/RemoteConfig';
+import {GetPoisIcons} from './examples/sdk/GetPoisIcons';
+import {SetCacheMaxAge} from './examples/sdk/SetCacheMaxAge';
+import {TiledBuilding} from './examples/sdk/TiledBuilding';
 import Wayfinding from './examples/wayfinding/Wayfinding';
 import NavigateToPoi from './examples/wayfinding/NavigateToPoi';
 import SelectPoi from './examples/wayfinding/SelectPoi';
