+++ example/src/examples/sdk/BuildingFullInfo.tsx
@@ -3,9 +3,9 @@ import {ScrollView} from 'react-native';
 import SitumPlugin from '@situm/react-native';
 import type {Building, Floor, Poi} from '@situm/react-native';
 
-import {SITUM_BUILDING_ID} from '../situm';
-import styles from './styles/styles';
-import {fetchBuilding} from './Utils/CommonFetchs';
+import {SITUM_BUILDING_ID} from '../../situm';
+import styles from '../styles/styles';
+import {fetchBuilding} from '../Utils/CommonFetchs';
 import {Card, Text} from 'react-native-paper';
 
 export const BuildingFullInfo = () => {
