+++ example/src/examples/sdk/InfoFromBuilding.tsx
@@ -2,9 +2,9 @@ import React, {useEffect, useState} from 'react';
 import {Text, ScrollView} from 'react-native';
 
 import SitumPlugin, {BuildingInfo} from '@situm/react-native';
-import {SITUM_BUILDING_ID} from '../situm';
-import styles from './styles/styles';
-import {fetchBuilding} from './Utils/CommonFetchs';
+import {SITUM_BUILDING_ID} from '../../situm';
+import styles from '../styles/styles';
+import {fetchBuilding} from '../Utils/CommonFetchs';
 import {Card} from 'react-native-paper';
 
 export const InfoFromBuilding = () => {
