+++ example/src/examples/sdk/SetCacheMaxAge.tsx
@@ -1,7 +1,7 @@
 import React, {useEffect, useState} from 'react';
 import {Text, ScrollView} from 'react-native';
 import SitumPlugin, {Building} from '@situm/react-native';
-import styles from './styles/styles';
+import styles from '../styles/styles';
 import {Button, Card} from 'react-native-paper';
 
 //You may use a few seconds in development and at least 30 minutes in production
